Change type from u32 to u8 to match IPv4 spec
Each segment of an ipv4 address can only be 0-255, so each of the
segments only need to be u8s.

If, however, this was done to not have to explain the difference between
u32 and u8, or if this would be distracting from the point about enums
trying to be made here, I would be fine reverting this change.