Rename files to better match mdbook generated sidebar numbering
I have had off-by-one errors when I'm talking to people and looking at
the file names, but they're looking at the sidebar in the generated
output, since the first section in a chapter doesn't get a number in the
sidebar, and the second section in a chapter gets named x.1, etc.

Just wait til I try to name cache invalidation.