Clarify that there isn't a separate "byte type"
The text indicated that there was separate a "byte type" in Rust.
However, based on the Rust reference, byte literals are simply a more
convenient way to represent `u8` numbers and arrays of `u8` numbers.
This change clarifies that there is no separate "byte type".

Also, the text suggested that only ASCII characters could be represented
with byte literals; however, non-ASCII 8-bit values can be represented
using backslash escapes in byte literals. This change changes the
wording to "8-bit values".