remove confusing section from chapter 2
This section was effectively re-teaching the concept of trivial changes causing a recompile, which should already be known this far in.

It was confusing for me, because I thought it was saying that the dependencies wouldn't be installed **until** a trivial change was made to `main.rs` (despite the section right above this saying the opposite).

It took me 10+ minutes to understand the mistake I'd made (including running the commands locally). Especially since the concept this is describing has already been mentioned earlier on, it doesn't seem worthwhile to mention it here.