Remove excess #s from code examples
I'm a rust beginner and working through your book as part of my learning process (I know this is a work in progress but still very useful to a beginner like me). 

I've been typing in the code examples as I go and noticed there were some parts of the code with extra #s in front of them which doesn't seem to be valid rust. They are also visible in the github pages version here http://rust-lang.github.io/book/ch05-01-method-syntax.html#Methods%20with%20More%20Arguments so this is a pull request to remove them.

Of course if they are supposed to be there then please just close this :)
Thanks for all the hard work you have put into this book!