Nit: explicitly mention the edition of the book
I'm never going to be able to visually distinguish between the first and second editions of the book, so going forward it will be nice to have some assurance that I'm linking people to the intended edition. It's also common practice for the covers or forewords of books to mention their edition, and the intro page is the closest thing the book has to either of those.