Update paragraph describing Listing 10-1
Some readers may complain about this type of step-by-step paragraph, but
I think they are nice.  I have tried to rewrite this one with special
attention on the choice of words.  In the original version, the
following terms were all used to refer to an individual element of the
list:

- integer
- item
- number
- value

I think this is a clarity issue.  Yes, all these mean the same thing
(hey, you didn't use `element`!), but using four terms in one paragraph
to refer to the same thing seems untidy.  I replaced all instances of
`item` and `value` with `number` to match the caption for Listing 10-1.
I specifically left in the one instance of `integers`, because it
identifies the data type and the same sentence refers to the variable,
`numbers`.

If this was documentation rather than a book, I would expect fairly
strict standardization of terms like these, but in a book format it
seems like the original text may be acceptable.  Consider this "food for
thought" more than anything else.  It kind of applies book-wide.

Also, this entire section refers to a `vec![1, 2, ...]` as a "list"
while the section on vectors didn't really use that term at all.  Is
there a special distinction between "list" and "vector" here? Did
you clarify somewhere that the term "list" was going to be
interchangeable with "vector"? I haven't read all those sections so
I just don't know.