Remove redundant note
The removed text is problematic for two reasons.

1. This is the first time the `..` range notation is used in the written
text in the entire book. The reader up to this point has not seen it
explained and it has only appeared in one code example where it
wasn't really mentioned.  Due to the choice of `[0..5]` as the
example here, it appears you are saying that the number before
the `..` is the starting index and the number after the `..` is the count.
This will get readers off on the wrong path until they make it a few
more paragraphs in.

2. The following paragraph explains the same thing about the
internal representation as a start and a length, so mentioning it here
is redundant.  It is better to keep this first paragraph simple.