Explain that string slice ranges are half-open
The discussion as a whole makes it clear that the ending index is exclusive,
but because of the way people new to the language will be reading this, I
think it is very important to make this explicit.

Some languages include a `[a..b]` syntax that looks very similar to Rust's,
but not all of them use the same rules.  Someone coming from Ruby, where
this notation on a string signifies a closed interval where both endpoints are
included may be very confused for a while right here.