Update ch17-02-trait-objects.md
This sentence was very hard to parse for me as a non-native english speaker. Maybe it's better this way? Also for a trait `Foo` one should imho not refer to a trait object as `Foo`, but give an example, e.g. `Box<Foo>`, which was introduced before. Just because `Foo` is a trait which is not a trait object (nitpick? Not sure, but I think this might lessen confusion).