Fixed a small inconsistency in code listings
I think `string2` being a string literal in Listing 10-21 doesn't make much sense and makes the example confusing while serving no educational purposes. Even more so when this is changed in Listing 10-24 and Listing 10-25.

While testing the examples, I copied the code in Listing 10-21 then later modified it to look like the code in Listing 10-25 and, surprisingly, it compiled. It took me a couple of minutes to figure out that the definition of `string2` was changed and not knowing yet about static lifetimes (explained later in the chapter) made it impossible for me to figure out why the program would behave like that.