Add the title of the link into its text
To make it clearer where links are going, eliminate entirely the
language of "Related" or the long verbose book titles. Instead, simply
prefix with "In the first/second edition" as needed, and in similar
style for other resources like the reference or the Unsafe book.

Add the target's title, i.e. chapter number, name, and sometimes
subsection, directly to the link copy.