Add blurb and code example
To improve immediate value and reduce discard-because-not-useful factor
upon arriving at a page which does not contain the content you were
expecting, add a small blurb taken/adapted from the newer source,
optionally accompanied by a code sample describing the feature.

However, due to the transitive nature of this resource, there must be no
content here that isn't available elsewhere.