Remove potential confusion of indexes and values
When encountering `enumerate()` for the first time, it's kind of confusing where all these items are coming from. The original example enumerates a list of integers. This may confuse people to think that the indexes are coming from the list, like a kind of `zip()`. If you instead use different types, this isn't possible.