Remove the comment about reusing variable names
We should not be spared to come up with different names, like "spaces_str" and "spaces_num". Reusing variable names for different types leads to assumption errors and it is difficult to track in bigger code bases. It has been considered a very bad practice for many years even in dynamic languages. Having this statement in the Rust book may "inspire" new programmers to implement this bad practice hence I have removed it.