Add Unit Tests for Rectangle Program
I previously read the first edition a few years ago. I'm now reading this second edition and coding along as I go with an emphasis on testing. I think readers should gain an appreciation of how unit tests fit in at the same time as understanding how to implement the code, as this is demanded by industry.