Add documentation for nested matching
I had no idea this feature existed until someone told me about it on Gitter. As far as I can tell it isn't documented in the book anywhere. If that's wrong, let me know :-)

I added this PR to the 2018 book because I assume it's too late & too big a change to add to the 2nd edition.