String variable should be mutable in Listing 4-6
The aim of Listing 4-6 is to demonstrate the a String passed by reference cannot be modified, however, as the example code is written the string could not be modified anyway because it is immutable. The proposed change is to make it mutable so that the only problem in the example code has to do with modifying a borrowed object.