clarify discussion of macros and namspaces
The existing text was rather unclear, for instance it might mislead a
person into the false understanding that ALL macros within a crate are
exported, regardless of export attribute (which is not mentioned).

If the text is going to make references to namespace differences
compared to functions and bringing them into scope and such, we might as
well have something clear and truly informative.