Mention more about `Cargo.lock` early on
I'm not 100% satisfied with this, but I think that it's important to mention what `Cargo.lock` is exactly when it's mentioned, rather than what's given. Because people will start committing code to VCS as soon as they start learning, it makes sense for them to understand why they might (or might not) commit `Cargo.lock` to their repos-- arguably, I'd say that it makes sense to start out by saying "if you're not sure what to do, commit `Cargo.lock`" just under the premise that most intro users will be building binaries, not libraries.

At least for now, clarifying `cargo update` seems reasonable.