use lifetimes 's and 'c in subtyped lifetimes
In the subtyped lifetimes section there is a paragraph where the names of lifetimes jump form `'s`  and `'c` to `'a` and '`b'. This shift in names makes following the example hard as it requires to constantly translate between `'a`, `'b'`, `'s` and `'c` - all being single letter abbreviations already making it even harder.