ch15-02-deref: Improve explanation on immut-to-mut
On a hypothetical immut-to-mut deref, the resulting mut ref should be
unique. The aliasing rules say a mutable reference should be the only
reference, so the initial immutable reference would have to be the only
one, which is not enforced. Change the phrasing so it expresses this
fact more clearly.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>