break semicolon removal
In this code example we use `break` to return the result of the expression `counter * 2`. The Rust compiler is fine with or without ending this line with a semicolon. But it seems more consistent to Not use the semicolon since the absence of the   ;   declares that this line is not a statement and this block will be returning a value.