Update ch01-02-hello-world.md
I’d like to suggest two changes to the Hello World example.

1) On Windows, enter the command .\main.exe instead of ./main:
This is using a Unix idiom rather than a generic Windows idiom for how a program is executed. It works, but it ain’t the Windows way. 
It should be changed to:
On Windows, enter the command main instead of ./main:

> rustc main.rs
> main

2) The term “print to” is really a left-over from another era of I/O devices. Sadly, it's been incorporated into the name of the "println" command, but it is worth pointing out at least once what it really means in this circumstance, which is to display something on the screen.

Regardless of your operating system, the string Hello, world! should print to the terminal.
could become:
Regardless of your operating system, the string Hello, world! should display on (“print to”) the terminal.

Kurt