Clarify Copy linguistics
A type as *being* "Copy" as an adjective collides with the common sense word "copy" being either a verb or a noun. Something is a copy of, or you copy from something to something else. The documentation should (I feel) reflect this not to confuse readers (as I was). I'm not sure whether the "ˋCopyˋable" semantics is congruent with Rust traits, having no experience with it, but I think it adds to the readability of this section.