Make sentence in ch17-02 clearer
Move the word "instead" within a sentence.

Original sentence:

The width, height, and label fields on Button will differ from the fields on other components, such as a TextField type, that might have those fields plus a placeholder field instead.

On my first reading of this sentence, I interpreted it as follows:

The width, height, and label fields on Button will differ from the fields on other components, such as a TextField type. A TextField type might have the same fields as a Button. Or another option is that a TextField has a placeholder field.

I propose moving the "instead" within the sentence, to read as follows:

The width, height, and label fields on Button will differ from the fields on other components, such as a TextField type, that might instead have those fields plus a placeholder field.

In this way it's clear that the "instead" refers to all of (those fields plus a placeholder field), and not just to (plus a placeholder field).