Proposed minor rewrite of the earlier parts of Chapter 7 that includes a slightly expanded overview of the Rust ecosystem before diving into the details of it.
There are some technical details I am unsure of.  If this is found to be
useful, someone needs to review the edits for technical accuracy.

(Specifically, it is my understaning that only Cargo udnerstands the
nature of crates and that it passes approrpriate commands lines to the
Rust compiler to locate the files.  Also, it is my understanding that
a module can be spread across multiple source files.)