ch04-01 add brief explanation of stack frames
Occasionally, newcomers will take the stack-of-plates analogy too
literally and become confused about why it's possible to access a
variable that's not at the top of the stack. Add a brief explanation
about stack frames that is hopefully good enough to help them
search for more details if they need them.