Update ch03-02-data-types.md
This tripped me up, because the compiler produced an error that pointed me to the generic, and only the generic. 

When I drilled down to the definition for the generic with some frantic, frustrated Googling though, it highlighted to me that the function I wanted, u8::wrapping_add(a, b), existed as a static function of the type, u8. Maybe this breadcrumb will help.