removed incorrect info about computer architecture
I've removed the bit about 32 bit integers sometimes being faster on 64 bit machines, which is only true when handling so much memory that the natural processing speed of the machine, 64 bits in this case, is over shadowed by physical limitations. In its place I've offered a brief introduction to the word size of the machine and the trade-offs between efficiency and space.