Make need to open new shell post-install clearer
The current text makes it easy to miss the fact that you need to open a new shell after the install for the new path variables to be usable. To make it clearer, add a discrete statement to open a new shell separate from the session you ran the install in. Then, when it runs through a list of troubleshooting steps, reiterate this fact so that the user has little chance of missing it. This will prevent vain troubleshooting from users who missed the "new shell" part and prevent unneeded forum posts.