Reflect the addition of the -p flag in Cargo 1.56 in chapter 14
In Cargo 1.56 (2021-10-21), support for publishing a package from a Cargo workspace
was added to the `cargo publish` command.

The book however still mentioned that it wasn't possible to publish a package
unless you navigated into the respective package directory and published from
there.

This commit fixes the text which mentioned it isn't possible, and the text
was instead changed to be an example of how the flag can be used (similar to
the `cargo test -p [spec]` above it).

The manual-regeneration comment was taken from Chapter 14.2 (Publishing a Crate
to Crates.io) as the example was similar to the publish a crate example there.