Use concrete names for concrete lifetimes in ch10-03
At the start of chapter 10-03, the concept of a "lifetime" is introduced, by talking about the concrete lifetimes of values. These lifetimes are referred to using `'a` and `'b`. This can be confusing, since these names are usually used for generic lifetimes. The difference between generic lifetimes and concrete lifetimes may not be obvious, so don't make it less obvious by using generic names for the concrete lifetimes.