Update ch10-03-lifetime-syntax.md
Lifetime parameters refer to the borrowed values not the function parameters. This distinction is important in cases where the actual parameters (the references) are (conceptually) stored only on the stack (in several examples of the chapter) -- in contrast to the actual values referred to which are created (and stored) in the caller's scope.