Update appendix-02-operators.md
Provide inclusion of the operators `!` and `?` for their use on auto traits. "Keywords" and "Operators" serve as a central index the documentation and therefore must be comprehensive. In particular, documentation for these symbols appearance in code on traits cannot be easily discovered since they are overloaded and heavily used in other circumstances, especially `!`.