Add missing subject to sentence about performance
I can't tell if this is the sort of change you accept or not, but I
figured it won't hurt to submit a PR. I searched for issues and PRs that
mentioned this problem in chapter10.md, but I didn't find any. That
said, when I build the book, I see that the problem must've already been
addressed elsewhere.

If you have to close this because reasons, I won't be offended!