Update ch07-01-packages-and-crates.md
I think it should be "A package can contain as many library crates as you like, but at most only one
binary crate. A package must contain at least one crate, whether that’s a
library or binary crate." from my understanding, and it would be a great misunderstanding to the practice.