Add a missing line generated by 'cargo new'
When running `cargo new hello_cargo` in the command line, one of the lines automatically generated in the Cargo.toml file of the `hello_cargo` project is a comment regarding more keys and their definitions. This comment is currently missing in the book chapter, and it probably should be added there.