Change "Under the covers" to "Under the hood"
In chapter 4.1, the authors write "Take a look at Figure 4-1 to see what is happening to String *under the covers.*"  Isn't the the common figure of speech "under the hood"?

"Under the hood"  if a reference to a car's bonnet (or hood), which conceals the complex inner workings of the engine. "Under the covers"  seems like a reference to something else.