Refactor: line 89-90 - let vars and global scope
Change made to make it clear that let variables cannot be declared in the global scope. This small addition makes it more explicitly clear. It's one of the questions asked in the follow along quiz version of the book and while following along it wasn't made clear until a question was asked. I think having it explicitly stated in the book will be helpful.