Include rectangle rotation for fn can_hold
In the example used for function can_hold the test to check is to see if the first rectangle has greater width than the second rectangle and greater height than the second rectangle. My proposal is to also allow the case that the second rectangle fits by "rotating" the rectangle so to speak. An example is if rectangle a has dimensions 10x5 and rectangle b has dimensions 4x6, then rectangle b would fit into rectangle a with this proposed change, but would not with the current function.