Remove awkward phrase about types of strings.
"Not just one of these" is unclear and confusing. The purpose of the sentence is to say that when people say "string", they could either mean String or &str. The meaning is conveyed perfectly well without the phrase "not just one of these", and that phrase makes the sentence significantly more difficult to read.