Edit a paragraph in ch02-00
The `trim` method on a `String` instance will eliminate any whitespace
at the beginning and end. This is important because to be able to
compare the string to the `u32`, the string can only contain numerical
data.

In the previous version, the dependant clause "can only contain
numerical data" described the `u32`. The revised paragraph makes it
clear that we are trimming the string so that it contains numerical data
only.

Signed-off-by: Jorge Troncoso <jatron@users.noreply.github.com>