Fix variable mutability in code example
The code example in the book had a mistake where the variable `x` was declared as immutable (`let x = 5;`) instead of mutable. This mistake was misleading to learners as it implied that the variable could not be modified. To rectify this issue, the commit corrects the variable mutability by adding the `mut` keyword (`let mut x = 5;`), enabling the modification of `x` later in the code.