Emit semantic HTML instead of blockquotes for notes
Use the newly-introduced `mdbook-trpl-note` to render accessible and
semantic HTML for the notes/callouts in the book. This currently applies
to the nostarch version as well, so a follow-on change will update the
build process to undo that. Add CSS so the new note rendering matches
the existing rendering of blockquotes. (We might want to change that at
some point, but for now the goal is to maximizing continuity.)