Do not apply semantic note handling to nostarch
Introduce a `book.toml` in the `nostarch` directory, which matches the
existing `book.toml` except for specifying the correct relative input
and output directories and disabling the `trpl-note` preprocessor. This
also lets us simplify the invocation in the `nostarch.sh` script a bit,
by providing the configuration directly in the file.