Ch13: Clarify closure capture types
Most places in this chapter are explicit about what type of capture is
occurring ((im)mutable reference or ownership), and this change updates
two more places to follow that standard. I've left the type of capture
unspecified in the description of listing 13-16 because it is just a
brief overview, and the capture type is now in the full breakdown
immediately after the listing.