Ch. 17: Drop material about `retry`, since it is a hazard
As described in the parent commit, what I have here technically works
but is adjacent to an area of the language which does *not* yet work as
well as we might hope: closures which involve async. Since even the
smallest refactor here (extracting the future *or* the closure to a
standalone binding) breaks the code, best to remove it entirely in favor
of some other example of composing futures together.

backout of commit d5ed8de41d5e5935d7a384dc21af59f77c8548cf