Update ch05-03-method-syntax.md
Instead of writing
"In `main`, when we follow `rect1.width` with parentheses, Rust knows we mean the method `width`. When we don’t use  parentheses, Rust knows we mean the field `width`.

adding Paranthesis to width() makes it more understandable
"In `main`, when we follow `rect1.width()` with parentheses, Rust knows we mean the method `width()`. When we don’t use  parentheses, Rust knows we mean the field `width`.