infra: *some* file paths can also contain `file:///home`
Rust recently began printing error messages which include the full path
to the source of the file with an error in it; when that path is in the
standard library, that includes the user's home directory (or similar).
A previous change (b9241f6) added support for those paths in the output
files; this one adds support to the local file paths linter.

Specifically allow exactly and only paths in the home directory which
include `.rustup` or `.cargo`.