WIP: infra: implement auto-formatting GitHub Action
Introduce a GitHub Action which allows us to add a comment `!format` to
a given PR to have its contents automatically formatted with dprint.
This lets contributors not worry about our exact formatting guidelines
and focus instead on contributing. People who *want* to go the extra
mile can, and we document local autoformatting setup in CONTRIBUTING.md,
but this way online-only contributions can “just work”.

WIP because this is one alternative; we may also:

- Use one of the existing Rust infra bots.
- Do this but include merge capability so we don’t have to do an extra
  dance when merging, just `!format-and-merge` and have it do both.
- Nothing!