Remove extraneous `use` statement (#4193)
In Listing 14-7 of Chapter 14, the `use add_one;` statement is redundant
and can be removed. Before the Rust 2018 edition, this statement was
required to bring the `add_one` crate into scope, so it likely got
missed during refactorings from that transition.

By removing this line, the code becomes more concise and reflects the
current best practices for using external crates in Rust.

Issue #4186

---------

Co-authored-by: Chris Krycho <hello@chriskrycho.com>