Fix 20.1 Miri unsafe code section
Fixes #4338 Wrong concept in subsection "Using Miri to check unsafe code"
of section "20.1. Unsafe Rust"

This is a followup to PR #4062

This PR adds a working example of using Miri where the code will compile
and run without warnings or errors, but with undefined behavior while it
runs. After running it through Miri though it correctly identifies the
data race.

Feedback welcome on matching the text to the books style. Or if anyone
wants to collaborate on other examples of Miri output.