Update the scraper dependency
I'm trying to get derive_more 2.0 (a crate I'm author of) on the
rust-playground. This involves getting the crates that pull in
derive_more to update their dependencies. This crate is the leaf crate
that pulls derive_more in. So hereby a PR to update the next one in the
chain.

Related:
- https://github.com/rust-lang/rust-playground/issues/1164#issuecomment-3020253643
- https://github.com/rust-scraper/scraper/pull/257