Avoid disparaging language
The word "dogmatic" has negative assocations with "irrational" and "unwilling to change ones mind".  The word "opinionated" on the other hand is more neutral and even positive in many situations. This is more fitting, because we're not trying to argue that the approaches of other languages are inherently worse than Rust's.